+++ src/Application/Command/Type/UpdateHandler.php
@@ -53,10 +53,15 @@ class UpdateHandler
     /**
      * @param Update $update
      *
+     * @throws PackageNotRequiredException
      * @throws TypeAlreadyExistsException
      */
     public function handle(Update $update)
     {
+        if(!$update->isPackageRequired && $update->isPaymentRequired) {
+            throw new PackageNotRequiredException();
+        }
+        
         $type = $update->type;
         $type->update(
             $update->rank,
@@ -67,7 +72,8 @@ class UpdateHandler
             $update->canRemoveMeeting,
             $update->areAllSheetParticipantsAssignedToMeeting,
             $update->canScanParticipant,
-            $update->isPackageRequired
+            $update->isPackageRequired,
+            $update->isPaymentRequired
         );
         $type->setHidden($update->hidden);