+++ src/Ui/Bundle/AdminBundle/Controller/Type/CreateAction.php
@@ -15,6 +15,7 @@ use Proximum\Vimeet\Application\Adapter\CommandBusInterface;
 use Proximum\Vimeet\Application\Adapter\RouterInterface;
 use Proximum\Vimeet\Application\Adapter\TranslatorInterface;
 use Proximum\Vimeet\Application\Command\Type\Create;
+use Proximum\Vimeet\Application\Command\Type\PackageNotRequiredException;
 use Proximum\Vimeet\Application\Exception\Type\TypeAlreadyExistsException;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\Type\TypeCreateType;
@@ -100,6 +101,10 @@ class CreateAction
                 foreach ($typeAlreadyExistsException->getLocales() as $locale) {
                     $form->get('translations')->get($locale)->get('title')->addError($error);
                 }
+            } catch (PackageNotRequiredException $packageNotRequiredException) {
+                $errorPayment = new FormError($this->translator->trans('admin.type.no_required_package'));
+
+                $form->get('isPaymentRequired')->addError($errorPayment);
             }
         }